src/requirements.txt
# Core LiveKit dependencies
livekit>=0.17.0
livekit-agents>=0.10.0
livekit-plugins-deepgram>=0.6.0
livekit-plugins-openai>=0.7.0
livekit-plugins-silero>=0.6.0

# HTTP and async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Audio processing (M2 optimized)
numpy>=1.24.0
pyaudio>=0.2.13

# Speech APIs
deepgram-sdk>=3.0.0
openai>=1.0.0

# Environment configuration
python-dotenv>=1.0.0

# Additional utilities
websockets>=12.0